scanner/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiohttp==3.9.3
httpx==0.26.0
requests==2.31.0
dnspython==2.4.2
cryptography==41.0.7
beautifulsoup4==4.12.3
tldextract==5.1.1
pydantic==2.5.3
python-dotenv==1.0.0
loguru==0.7.2
slowapi==0.1.9
